Rename 'raw' module to 'client' (#613)
The 'http::raw' module was initially intended as a home for oneshot
request functions. As this has now gone away and been replaced by a full
client, the module itself should be renamed to 'client' to match it.
