+++ src/builder/edit_profile.rs
@@ -29,7 +29,7 @@ impl EditProfile {
     ///
     /// # impl EventHandler for Handler {
     ///    # fn message(&self, context: Context, _: Message) {
-    ///         use serenity::{http::raw, utils};
+    ///         use serenity::utils;
     ///
     ///         // assuming a `context` has been bound
     ///
