+++ src/http/mod.rs
@@ -23,16 +23,16 @@
 //! [`Client`]: ../client/struct.Client.html
 //! [model]: ../model/index.html
 
+pub mod client;
 pub mod ratelimiting;
-pub mod raw;
 pub mod request;
 pub mod routing;
 
 mod error;
 
 pub use reqwest::StatusCode;
+pub use self::client::*;
 pub use self::error::Error as HttpError;
-pub use self::raw::*;
 
 use reqwest::{
     Method,
@@ -64,7 +64,7 @@ use crate::client::Context;
 /// [`Context`], you can pass a tuple of `(CacheRwLock, Http)`.
 ///
 /// [`CacheRwLock`]: ../cache/struct.CacheRwLock.html
-/// [`Http`]: raw/struct.Http.html
+/// [`Http`]: client/struct.Http.html
 /// [`Context`]: ../client/struct.Context.html
 pub trait CacheHttp {
     #[cfg(feature = "http")]
